pyproject.toml
[project]
name = "dicom-cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "libsixel-python>=0.5.0",
    "numpy>=2.3.0",
    "pillow>=11.2.1",
    "pydicom>=3.0.1",
    "textual>=3.3.0",
    "textual-image>=0.8.2",
]
